requirements.txt
# AE_LSTM project dependencies

# Core dependencies
numpy>=1.19.0
scipy>=1.6.0
pandas>=1.2.0
scikit-learn>=0.24.0

# Deep learning
torch>=1.9.0
torchvision>=0.10.0

# Data handling
h5py>=3.3.0
xarray>=0.19.0
netCDF4>=1.5.7

# Configuration and experiment management
hydra-core>=1.1.0
omegaconf>=2.1.1
optuna>=2.9.0
tensorboard>=2.6.0
wandb>=0.12.0  # Optional: for Weights & Biases integration

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0

# Utilities
tqdm>=4.61.0
PyYAML>=6.0
joblib>=1.0.1

# Optional: For advanced visualizations
umap-learn>=0.5.1
